+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - `deepresonsance` support
+- `additional lanterns` support
 
 ## [0.5.2] - 2023-06-24